services/idp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyJWT[crypto]==2.9.0
cryptography==43.0.0
orjson==3.10.6

